Fix section 3 bug
The author of the quote - thor - was moving around.
Fix it by wrapping all content in section 3 in a flex
container. Set it with a width of 830px, put that in another
flex container, and centered the items. This allowed me to have
the author of the quote at the right side, and the text of the
quote in the left side, and them not moving in different ways.